Add files via upload
This code uses the “lxml” and “pandas” libraries in Python to parse an XML file named "compiler.xml" and extract certain data from it. The extracted data is then written to an Excel file.
The code starts by importing the necessary libraries, “lxml” and pandas.
It then uses the “etree.parse()” function to parse the "compiler.xml" file and get the root element of the XML file. It initializes an empty list called "bookinfo" to store the extracted data.
It then uses a for loop to iterate through all elements in the XML file with the tag "book". For each "book" element, it creates an empty dictionary named "book_dict" and adds the following information to the dictionary:
the value of the attribute "id" of the "book" element, stored under the key "Book ID"
the text value of the "author" element within the "book" element, stored under the key "Author"
the text value of the "title" element within the "book" element, stored under the key "Title"
the text value of the "genre" element within the "book" element, stored under the key "Genre"
the text value of the "price" element within the "book" element, stored under the key "Price"
the text value of the "publish_date" element within the "book" element, stored under the key "Publish Date"
the text value of the "description" element within the "book" element, stored under the key "Description"
After processing all the "book" elements, it appends the "book_dict" to the "bookinfo" list.
Then it creates a “dataframe” from the "bookinfo" list using the pandas library, assigns it to variable 'data' and prints it.
Lastly, it writes the 'data' to an Excel file named”200901082_assign_3”